simplified timer calculation and fixed "remaining" to the new agent timer scheme Farcry still not running stable

+++ src/main/java/de/flashheart/rlg/commander/mechanics/ScheduledGame.java
@@ -36,8 +36,9 @@ public abstract class ScheduledGame extends Game {
 
     protected void deleteJob(JobKey jobKey) {
         if (jobKey == null) return;
-        log.debug("deleting Job {}", jobKey);
         try {
+            if (scheduler.checkExists(jobKey)) return;
+            log.debug("deleting Job {}", jobKey);
             scheduler.interrupt(jobKey);
             scheduler.deleteJob(jobKey);
             jobs.remove(jobKey);
